refactor(executor): split wiring from domain core, dedupe, name actors
Phase 0b of RFC-0002 prep — behavior-preserving extractions in the files
the data-plane phases rewrite:

- start.go: StartExecutor, Manager/cache wiring, executorControllers,
  executorAPIServer, runAdoptCleanup, bindAddr move verbatim out of
  executor.go, which now holds only the Executor domain core (request
  multiplexer + service creation).
- withSpecializationTimeout replaces the twice-copy-pasted
  specialization-timeout block in serveCreateFuncServices.
- gpm.go: the service() actor's switch arms become named methods
  (handleGetPool/handleCleanupPool/handleGetEnvPools);
  AdoptExistingResources splits into adoptPools /
  adoptPerImagePoolDeployments / adoptSpecializedPods. One bug fixed in
  the move (flagged): CleanupOldExecutorObjects logged the last
  individual err instead of the joined errs.
- POD_READY_TIMEOUT is parsed once in MakeGenericPoolManager instead of
  on every pool creation; MakeGenericPool takes it as a parameter.
- api.go: the duplicated poolmgr/newdeploy cache-validity dance
  collapses into serveFromCache (failOnCacheError preserves the poolmgr
  error semantics).
- Anonymous goroutines named: executorControllers readiness wait →
  markSyncedWhenCachesWarm; executor client tap batch → flushTaps.

TestPoolCacheRequests run 50x with -race: one failure observed, matching
the documented pre-existing flake (fscache is untouched by this diff,
confirmed via git diff main -- pkg/executor/fscache = empty).
